+++ scripts/systems/pf1.js
@@ -5,6 +5,11 @@ export class EnginePF1 extends Engine {
 
   constructor() {
     super();
+  }
+
+  init() {
+    super.init();
+
     game.settings.register(Stealthy.MODULE_ID, 'spotTake10', {
       scope: 'world',
       config: false,
@@ -19,17 +24,6 @@ export class EnginePF1 extends Engine {
       config: true,
       type: Number,
       default: -999,
-    });
-
-    Hooks.once('ready', () => {
-      const offset = game.settings.get(Stealthy.MODULE_ID, 'passiveSpotOffset');
-      if (offset === -999) {
-        game.settings.set(
-          Stealthy.MODULE_ID,
-          'passiveSpotOffset',
-          game.settings.get(Stealthy.MODULE_ID, 'spotTake10') ? 10 : -99
-        )
-      }
     });
 
     Hooks.on('pf1ActorRollSkill', async (actor, message, skill) => {
@@ -47,6 +41,19 @@ export class EnginePF1 extends Engine {
         .insertBefore($('[name="stealthy.passiveSpotOffset"]')
           .parents('div.form-group:first'));
     });
+  }
+
+  ready() {
+    super.init();
+
+    const offset = game.settings.get(Stealthy.MODULE_ID, 'passiveSpotOffset');
+    if (offset === -999) {
+      game.settings.set(
+        Stealthy.MODULE_ID,
+        'passiveSpotOffset',
+        game.settings.get(Stealthy.MODULE_ID, 'spotTake10') ? 10 : -99
+      );
+    }
   }
 
   async setValueInEffect(flag, skill, value, sourceEffect) {
@@ -96,23 +103,23 @@ export class EnginePF1 extends Engine {
     super.rollPerception();
   }
 
-  findHiddenEffect(actor) {
+  findStealthEffect(actor) {
     const beforeV11 = Math.floor(game.version) < 11;
     return actor?.items.find((i) => i.system.active && (beforeV11 ? i.label : i.name) === 'Hidden');
   }
 
-  findSpotEffect(actor) {
+  findPerceptionEffect(actor) {
     const beforeV11 = Math.floor(game.version) < 11;
     return actor?.items.find((i) => i.system.active && (beforeV11 ? i.label : i.name) === 'Spot');
   }
 
-  makeHiddenEffectMaker(name) {
-    Stealthy.log('PF1.makeHiddenEffectMaker not used in PF1');
+  makeStealthEffectMaker(name) {
+    Stealthy.log('PF1.makeStealthEffectMaker not used in PF1');
     return (flag, source) => null;
   }
 
-  makeSpotEffectMaker(name) {
-    Stealthy.log('PF1.makeSpotEffectMaker not used in PF1');
+  makePerceptionEffectMaker(name) {
+    Stealthy.log('PF1.makePerceptionEffectMaker not used in PF1');
     return (flag, source) => null;
   }
 
@@ -121,8 +128,8 @@ export class EnginePF1 extends Engine {
     return null;
   }
 
-  async updateOrCreateHiddenEffect(actor, flag) {
-    let hidden = this.findHiddenEffect(actor);
+  async updateOrCreateStealthEffect(actor, flag) {
+    let hidden = this.findStealthEffect(actor);
     const beforeV11 = Math.floor(game.version) < 11;
     hidden ??= actor?.items.find((i) => (beforeV11 ? i.label : i.name) === 'Hidden');
     if (!hidden) {
@@ -150,8 +157,8 @@ export class EnginePF1 extends Engine {
     stealthy.socket.executeForEveryone('RefreshPerception');
   }
 
-  async updateOrCreateSpotEffect(actor, flag) {
-    let spot = this.findSpotEffect(actor);
+  async updateOrCreatePerceptionEffect(actor, flag) {
+    let spot = this.findPerceptionEffect(actor);
 
     // PF1 buffs can be disabled, if so, look for one already on the actor
     const beforeV11 = Math.floor(game.version) < 11;
@@ -193,6 +200,7 @@ Hooks.once('init', () => {
     const systemEngine = new EnginePF1();
     if (systemEngine) {
       window[Stealthy.MODULE_ID] = new Stealthy(systemEngine);
+      systemEngine.init();
     }
   }
 });